refactor(mentor): CSS 색상 변수 누락 건 수정 및 불필요한 코드 분기로 source.js 삭제
