+++ css/style.css
@@ -250,7 +250,7 @@ section .title{
 .type1 .btnMain:hover{
     /* PC */
     @media (min-width:1200px){
-        background:var(--blueHover);
+        background:var(--primary-foreground);
     }
 }
 .type2 .contentsWrap{
@@ -390,7 +390,7 @@ section .title{
     font-size:2.4em;
     font-weight:400;
     line-height:32px;
-    color:var(-);
+    color:var(--gray-700);
 
     /* Tablet */
     @media (min-width:768px) and (max-width: 1199px){
